refactor: consolidate simplified routers into main routers
- Replace original query-analyzer, schema-selector, and sql-builder with simplified versions
- Reduce AI structured output fields by 50-70% for faster generation
- Update workflow orchestrators to use new simplified schemas
- Fix all TypeScript type errors after consolidation
- Remove redundant -simplified.ts files

This optimization maintains precision while significantly reducing AI generation complexity and improving response times.
