--- src/server/api/routers/query-analyzer-simplified.ts
@@ -1,171 +1,0 @@
-import { createOpenAI } from "@ai-sdk/openai";
-import { generateObject } from "ai";
-import { z } from "zod";
-
-import { env } from "@/env";
-import { createTRPCRouter, publicProcedure } from "@/server/api/trpc";
-import { cacheManager } from "@/server/lib/cache-manager";
-
-/**
- * Simplified Query Analyzer - Optimized for faster AI generation
- *
- * Key improvements:
- * - 53% fewer fields in structured output
- * - 70% simpler structure (no deep nesting)
- * - Focused on essential routing decisions
- */
-
-// Simplified schema - only essential fields
-export const SimpleQueryAnalysisSchema = z.object({
-	routing: z.object({
-		strategy: z.enum(["sql_only", "vector_only", "hybrid", "rejected"]),
-		confidence: z.number().min(0).max(1),
-	}),
-	sqlTables: z.array(z.string()).optional(),
-	vectorQueries: z
-		.array(
-			z.object({
-				table: z.string(),
-				field: z.string(),
-				query: z.string(),
-			}),
-		)
-		.optional(),
-	feasible: z.boolean(),
-	reason: z.string().optional(),
-});
-
-export type SimpleQueryAnalysis = z.infer<typeof SimpleQueryAnalysisSchema>;
-
-export const queryAnalyzerSimplifiedRouter = createTRPCRouter({
-	analyzeSimple: publicProcedure
-		.input(
-			z.object({
-				query: z.string().min(1),
-				databaseSchema: z.string(),
-				vectorizedFields: z.record(z.array(z.string())).optional(),
-			}),
-		)
-		.mutation(async ({ input }) => {
-			const startTime = Date.now();
-			console.log("[SimpleQueryAnalyzer] 开始分析查询:", input.query);
-
-			// Generate cache key
-			const cacheKey = cacheManager.generateCacheKey("query-simple", {
-				query: input.query,
-				schemaHash: input.databaseSchema.substring(0, 100),
-			});
-
-			// Check cache
-			const cached = await cacheManager.getQueryAnalysis(cacheKey);
-			if (cached) {
-				console.log("[SimpleQueryAnalyzer] 使用缓存结果");
-				// Convert to simple format
-				const simpleAnalysis: SimpleQueryAnalysis = {
-					routing: {
-						strategy: cached.routing.strategy,
-						confidence: cached.routing.confidence,
-					},
-					sqlTables: cached.sqlConfig?.tables,
-					vectorQueries: cached.vectorConfig?.queries.map((q: any) => ({
-						table: q.table,
-						field: q.fields[0] || "",
-						query: q.searchText,
-					})),
-					feasible: cached.feasibility.isFeasible,
-					reason: cached.feasibility.reason,
-				};
-				return {
-					success: true,
-					analysis: simpleAnalysis,
-					executionTime: Date.now() - startTime,
-					cached: true,
-				};
-			}
-
-			try {
-				const openai = createOpenAI({
-					apiKey: env.AIHUBMIX_API_KEY,
-					baseURL: env.AIHUBMIX_BASE_URL,
-				});
-
-				// Parse schema to extract table names for context
-				const schema = JSON.parse(input.databaseSchema);
-				const tableNames = Object.keys(schema);
-				const vectorizedInfo = input.vectorizedFields
-					? `向量化字段: ${JSON.stringify(input.vectorizedFields)}`
-					: "无向量化字段";
-
-				const systemPrompt = `你是查询分析专家。快速分析查询并返回最简结果。
-
-数据库表: ${tableNames.join(", ")}
-${vectorizedInfo}
-
-路由规则:
-- sql_only: 可用SQL解决（包括LIKE模糊搜索）
-- vector_only: 纯语义搜索（找相似概念）
-- hybrid: 需要SQL条件+语义理解
-- rejected: 不可执行
-
-只返回必要信息:
-1. routing.strategy 和 confidence (0-1)
-2. sqlTables: 需要的表名列表
-3. vectorQueries: 向量搜索配置(如需要)
-4. feasible: 是否可行
-5. reason: 简短说明(可选)`;
-
-				const { object: analysis } = await generateObject({
-					model: openai("gpt-4o-mini"), // Use faster model for simple schema
-					system: systemPrompt,
-					prompt: `分析: "${input.query}"`,
-					schema: SimpleQueryAnalysisSchema,
-					temperature: 0.1,
-				});
-
-				// Don't cache simplified results - they're derived from full results
-
-				console.log("[SimpleQueryAnalyzer] 分析完成:", {
-					strategy: analysis.routing.strategy,
-					confidence: analysis.routing.confidence,
-					tablesCount: analysis.sqlTables?.length || 0,
-					executionTime: Date.now() - startTime,
-				});
-
-				return {
-					success: true,
-					analysis,
-					executionTime: Date.now() - startTime,
-					cached: false,
-				};
-			} catch (error) {
-				console.error("[SimpleQueryAnalyzer] 分析错误:", error);
-				throw new Error("查询分析失败");
-			}
-		}),
-
-	// Compare with original analyzer
-	compareAnalyzers: publicProcedure
-		.input(z.object({ query: z.string() }))
-		.mutation(async ({ input }) => {
-			const schema = JSON.stringify({
-				companies: { id: "int", name: "text", country: "text", rating: "int" },
-				contacts: { id: "int", companyId: "int", email: "text" },
-			});
-
-			// Return a simple comparison without circular dependency
-			return {
-				query: input.query,
-				simple: {
-					time: 100, // Mock time
-					fields: 8,
-					strategy: "hybrid" as const,
-				},
-				original: {
-					time: 200, // Mock time
-					fields: 17,
-					strategy: "hybrid" as const,
-				},
-				speedup: "50.0%",
-			};
-		}),
-});
